agents/quantum_arbitrage_agent/quantum_optimizer.py
"""
Quantum Optimization Module
Implements quantum optimization algorithms for arbitrage path finding
"""

import numpy as np
from typing import List, Dict, Optional
from dwave.system import DWaveSampler, EmbeddingComposite
from dimod import Binary, BinaryQuadraticModel
import pennylane as qml

class QuantumPathOptimizer:
    """Quantum optimizer for finding optimal arbitrage paths."""
    
    def __init__(self, n_markets: int = 10):
        """Initialize quantum path optimizer.
        
        Args:
            n_markets: Number of markets to analyze
        """
        self.n_markets = n_markets
        
        # PennyLane device for quantum optimization
        self.n_qubits = self._calculate_required_qubits()
        self.dev = qml.device("default.qubit", wires=self.n_qubits)
        self.circuit = qml.QNode(self._create_qaoa_circuit, self.dev)
        
        # D-Wave sampler for quantum annealing
        self.sampler = EmbeddingComposite(DWaveSampler())
        
    def _calculate_required_qubits(self) -> int:
        """Calculate number of qubits needed for the optimization."""
        # We need log2(n_markets) qubits to encode each market
        # and additional qubits for path constraints
        return int(np.ceil(np.log2(self.n_markets))) * 3
        
    def _create_qaoa_circuit(self, params, adjacency_matrix):
        """Create QAOA circuit for path optimization."""
        gamma, beta = params[0], params[1]
        
        # Initialize in superposition
        for i in range(self.n_qubits):
            qml.Hadamard(wires=i)
            
        # Problem unitary
        for i in range(self.n_markets):
            for j in range(self.n_markets):
                if i != j and adjacency_matrix[i,j] != 0:
                    qml.CNOT(wires=[i, j])
                    qml.RZ(gamma * adjacency_matrix[i,j], wires=j)
                    qml.CNOT(wires=[i, j])
                    
        # Mixing unitary
        for i in range(self.n_qubits):
            qml.RX(2 * beta, wires=i)
            
        return [qml.expval(qml.PauliZ(i)) for i in range(self.n_qubits)]
        
    def _create_qubo_model(self, 
                          price_matrix: np.ndarray,
                          min_profit: float) -> BinaryQuadraticModel:
        """Create QUBO model for D-Wave quantum annealing."""
        # Initialize QUBO model
        bqm = BinaryQuadraticModel('BINARY')
        
        # Add variables for each possible path segment
        vars = {}
        for i in range(self.n_markets):
            for j in range(self.n_markets):
                if i != j:
                    vars[f'x_{i}_{j}'] = Binary(f'x_{i}_{j}')
                    
        # Add constraints and objective
        # Path continuity constraints
        for i in range(self.n_markets):
            constraint = sum(vars[f'x_{i}_{j}'] for j in range(self.n_markets) if i != j) == 1
            bqm.add_constraint(constraint, label=f'continuity_{i}')
            
        # Profit objective
        objective = sum(price_matrix[i,j] * vars[f'x_{i}_{j}']
                       for i in range(self.n_markets)
                       for j in range(self.n_markets) if i != j)
        bqm.set_objective(objective)
        
        return bqm
        
    async def find_optimal_path(self,
                              price_matrix: np.ndarray,
                              min_profit: float = 0.01,
                              use_annealing: bool = True) -> Dict:
        """Find optimal arbitrage path using quantum optimization.
        
        Args:
            price_matrix: Matrix of exchange rates between markets
            min_profit: Minimum profit threshold
            use_annealing: Whether to use D-Wave annealing (True) or QAOA (False)
            
        Returns:
            Dictionary containing optimal path and expected profit
        """
        if use_annealing:
            # Use D-Wave quantum annealing
            bqm = self._create_qubo_model(price_matrix, min_profit)
            response = self.sampler.sample(bqm, num_reads=1000)
            
            # Get best solution
            best_solution = response.first.sample
            path = self._extract_path(best_solution)
            profit = self._calculate_profit(path, price_matrix)
            
        else:
            # Use QAOA
            # Initialize parameters for QAOA
            params = np.array([0.1, 0.1])
            
            # Run circuit
            result = self.circuit(params, price_matrix)
            path = self._extract_path_from_qaoa(result)
            profit = self._calculate_profit(path, price_matrix)
            
        return {
            'path': path,
            'profit': profit,
            'confidence': self._calculate_confidence(profit, min_profit)
        }
        
    def _extract_path(self, solution: Dict) -> List[int]:
        """Extract path from quantum solution."""
        path = []
        current = 0  # Start from first market
        
        for _ in range(self.n_markets):
            next_market = None
            for j in range(self.n_markets):
                if j != current and solution.get(f'x_{current}_{j}', 0) == 1:
                    next_market = j
                    break
            if next_market is None:
                break
            path.append(next_market)
            current = next_market
            
        return path
        
    def _calculate_profit(self, path: List[int], 
                         price_matrix: np.ndarray) -> float:
        """Calculate total profit for a given path."""
        if not path:
            return 0.0
            
        profit = 1.0
        current = 0
        for next_market in path:
            profit *= price_matrix[current, next_market]
            current = next_market
            
        return profit - 1.0  # Convert to percentage
        
    def _calculate_confidence(self, profit: float, 
                            min_profit: float) -> float:
        """Calculate confidence score for the solution."""
        if profit <= min_profit:
            return 0.0
        return min(1.0, profit / (2 * min_profit))
